Rewrite random_bit_generator completely, and include only one algo
One algorithm with more powerful randomization should be better than
different ones with less randomness capability. The current
algorithm is far better than previous ones, but somehow predictable.
However, the algorithm is *incomplete*.

The algorithm just randomly selects a bit out of and replaces a bit
of the seed and moves on. The selection is based on a fibonacci-like
series. The selector is an array of integers, in which, its next
number is calculated from the sum of the two previous numbers. This
is what the algorithm generally does; for more details, refer to the
actual written code.